punto 3

+++ Alquileres/alquileres/src/main/java/utn/backend/grupo128/alquileres/repositories/AlquilerRepository.java
@@ -3,5 +3,5 @@ package utn.backend.grupo128.alquileres.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import utn.backend.grupo128.alquileres.models.Alquiler;
 
-public interface AlquilerRepository extends JpaRepository<Alquiler, Long> {
+public interface AlquilerRepository extends JpaRepository<Alquiler, Integer> {
 }
